Workaround for lack of archives plugin at GitHub
GitHub does not support the Jekyll-archives plugin for generating
content based on the dates of the posts. i.e. create an archive. The
workaround uses a layout that checks the current directory of a page.
Since pages use a permalink of /year/month/post and a file in /2017/
with the archive-year layout will yield a directory of /2017/, it can be
used to group all posts of that year and thus create an archive for that
year.